# Boost Ntriples performance