--- a/lib/Literal/DateTime.php
+++ b/lib/Literal/DateTime.php
@@ -75,13 +75,13 @@ class DateTime extends Date
 
         // workaround for PHP < 7.2: if not used, script will throw "Undefined class constant 'ATOM'"
         if (\defined('\DateTimeInterface::ATOM')) {
-            $format = DateTimeInterface::ATOM;
+            $format = \DateTimeInterface::ATOM;
         } else {
             $format = \DateTime::ATOM;
         }
 
         // Convert DateTime objects into string
-        if ($value instanceof DateTimeInterface) {
+        if ($value instanceof \DateTimeInterface) {
             $atom = $value->format($format);
             $value = preg_replace('/[\+\-]00(\:?)00$/', 'Z', $atom);
         }