--- a/lib/Parser/Json.php
+++ b/lib/Parser/Json.php
@@ -128,10 +128,10 @@ class Json extends RdfPhp
      * @param string              $format  the format of the input data
      * @param string              $baseUri the base URI of the data being parsed
      *
+     * @return int The number of triples added to the graph
+     *
      * @throws Exception
      * @throws \EasyRdf\Exception
-     *
-     * @return int The number of triples added to the graph
      */
     public function parse($graph, $data, $format, $baseUri)
     {