--- a/lib/Parser/Rdfa.php
+++ b/lib/Parser/Rdfa.php
@@ -669,9 +669,9 @@ class Rdfa extends Parser
      * @param string $format  the format of the input data
      * @param string $baseUri the base URI of the data being parsed
      *
-     * @throws \EasyRdf\Exception
-     *
      * @return int The number of triples added to the graph
+     *
+     * @throws \EasyRdf\Exception
      */
     public function parse($graph, $data, $format, $baseUri)
     {