… .github/workflows/Tests.yml
… README.md
… composer.json
… examples/uk_postcode.php
… examples/wikidata_villages.php
… lib/Serialiser/Turtle.php
… lib/Sparql/Client.php
… lib/Sparql/Result.php
… lib/TypeMapper.php
… lib/Utils.php
… test/EasyRdf/HttpTest.php
… test/EasyRdf/Literal/DateTest.php
… test/EasyRdf/Literal/DateTimeTest.php
… test/EasyRdf/LiteralTest.php
… test/EasyRdf/ParserTest.php
… test/EasyRdf/ResourceTest.php
… test/EasyRdf/Serialiser/NtriplesTest.php
… test/EasyRdf/Serialiser/TurtleTest.php
… test/EasyRdf/SerialiserTest.php
… test/EasyRdf/Sparql/ClientTest.php
… test/EasyRdf/TypeMapperTest.php
… test/EasyRdf/UtilsTest.php
… test/TestHelper.php
… test/examples/HttpgetTest.php
… test/examples/SparqlqueryformTest.php
… test/examples/UkpostcodeTest.php
… test/examples/WikidataVillagesTest.php