Ok, so before the long potato, REALLY SORRY for the long potato.

So we have been having some performance issues with the Ntriples serializer which we are using in our Drupal website. The client inserted a long text and I can assume that it is also because of a not blazing fast server, but the thing is that the Ntriples serializer was VERY slow.

We tried to check for alternatives like hardf library as was initially suggested (and abandonded) at https://github.com/easyrdf/easyrdf/issues/353 but there were multiple issues. For starters, we gave up hardf because indeed, it is much faster than easyrdf in ntriples at least, but testing the extended table of characters as described in https://www.w3.org/TR/rdf-testcases/#ntrip_strings, we found out that it is not really doing the job right (It was working sweet though for normal text). That is because, according to the link above, the Ntriples format kinda supports everything, which is a huge problem to be honest.

So we came back to the effort of optimizing this library's class instead. I will not get into huge details of how it took 3 days to understand all edge cases between multibyte strings and different encodings and adapting these to PHP, I am going to stick with the fact that 1. almost all current functionality is intact, 2. performance is greatly improved.

1. **Almost** all functionality kept
Part of our tests were to check the behavior of all the basic ASCII characters, i.e. 0-255. For that purpose, I created a script to attach each character next to each other. What I found out though is that after serializing, I was getting the following:
```
"\u0000\u0001\u0002\u0003\u0004\u0005\u0006\u0007\u0008\t\n\u000B\u000C\r\u000E\u000F\u0010\u0011\u0012\u0013\u0014\u0015\u0016\u0017\u0018\u0019\u001A\u00
1B\u001C\u001D\u001E\u001F !\"#$%&'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_`abcdefghijklmnopqrstuvwxyz{|}~\u007F\u0080\u0081\u0082\u0083\u0084\u0085\u0086\u0087\u0088\u0089\u008A\u00
8B\u008C\u008D\u008E\u008F\u0090\u0091\u0092\u0093\u0094\u0095\u0096\u0097\u0098\u0099\u009A\u009B\u009C\u009D\u009E\u009F\u00A0\u00A1\u00A2\u00A3\u00A4\u00A5\u00A6\u00A7\u00A8\u00A9\u00AA\u00AB\u00A
C\u00AD\u00AE\u00AF\u00B0\u00B1\u00B2\u00B3\u00B4\u00B5\u00B6\u00B7\u00B8\u00B9\u00BA\u00BB\u00BC\u00BD\u00BE\u00BF\u00C0\u00C1\u00C3\u014C\u00AD\u00AE\u00AF\u00B0\u00B1\u00B2\u00B3\u00B4\u00B5\u00B6\u00B7\u00B8\u00B9\u00BA\u00BB\u00BC\u00BD\u00BE\u00BF\u00C0\u00C1\u00C3\u0145\u01C7\u0249\u02CB\u034D\u03CF\u0451\u04D3\u0555\u05D7\u0659\u06DB\u075D\u07DF\u18A2\u4965\u7A28\uAAEB\uDBAE\U00010C71\U0C\u00AD\u00AE\u00AF\u00B0\u00B1\u00B2\u00B3\u00B4\u00B5\u00B6\u00B7\u00B8\u00B9\u00BA\C\u00AD\u00AE\u00AF\u00B0\u00B1\u00B2\u00B3\u00B4\u00B5\u00B6\u00B7\u00B8\u00B9\u00BA\u00BB\u00BC\u00BD\u00BE\u00BF\u00C0\u00C1\u00C3\u0145\u01C7\u0249\u02CB\u034D\u03CF\u0451\u04D3\u0555\u05D7\u0659\u06DB\u075D\u07DF\u18A2\u4965\u7A28\uAAEB\uDBAE\U00010C71\U000F4D75\u00F6\u00F7\u00F8\u00F9\u00FA\u00FB\u00FC\u00FD\u00FE\u00FF"
```
Please, note, that after `\u00C1`, the sequence should continue with `C2`, `C3` until `FF` which does not. Please, correct me if I am wrong, but I can assume that the current implementation, can confuse sequences for multibytes if it randomly happens. If I am correct, this is a bug I had to fix here too.

2. Performance difference in the ticket:
I ran 3 tests. First test, was to serialize the characters 0 to 255 + the unicode characters `位於` repeated 100 times. The unicode characters where there to ensure that they were interpreted properly. The results are with the above bug for the old code. **_Please, note that in the new code, while the sequences that are confused by the old code are in the same string as the multibyte characters above, the end result produced is correct. The multi byte characters are interpreted correctly as well as the above sequences._**
For 258 characters, the improvement was not noticeable but quite a bit:
```
New code took 0.0076038837432861 seconds. Memory usage: 736440 bytes. Memory peak usage: 1451512 bytes.      
Old code took 0.61801195144653 seconds. Memory usage: 850392 bytes. Memory peak usage: 1451512 bytes.
Number of characters: 26200

~ 80 times faster
```
Then I raised the number of characters by repeating the above, not 100, but 1000 times. The results were much better:
```
Ntriples::serialise() took 0.079196929931641 seconds. Memory usage: 2149560 bytes. Memory peak usage: 3697808 bytes.
NtriplesOld::serialise() took 64.083306074142 seconds. Memory usage: 2955736 bytes. Memory peak usage: 3848944 bytes.
Number of characters: 262000

~800 times faster
```
And then I also tried to mix it with more normal characters because in my mind, this is more of the most normal scenario (to have simple text mixed with special charactesr) by using the following:
```php
str_repeat($test_string . str_repeat('abcdefghigklmnopqrstuvwxyz', 50), 100))
```
where `$test_string` is the 258 characters from above, and the results were:
```
Ntriples::serialise() took 0.051412105560303 seconds. Memory usage: 2981048 bytes. Memory peak usage: 3848944 bytes.
NtriplesOld::serialise() took 28.111695051193 seconds. Memory usage: 2136536 bytes. Memory peak usage: 3848944 bytes.
Number of characters: 156200
~ 550 times faster
```

I have added an extended documentation on why I am using this sequence + here is the script for the performance check though I don't know if it is needed.
```php
<?php

use EasyRdf\Serialiser\Ntriples;
use EasyRdf\Serialiser\NtriplesOld;

include 'vendor/autoload.php';

function serializeString(string $string): string {
    // Test the performance of the Ntriples::escapeString() method.
    $start = microtime(true);

    $ntriples = new Ntriples();
    $output = $ntriples->serialiseValue(['type' => 'literal', 'value' => $string]);

    $end = microtime(true);
    echo 'Ntriples::serialise() took ' . ($end - $start) . ' seconds. Memory usage: ' . memory_get_usage() . ' bytes. Memory peak usage: ' . memory_get_peak_usage() . ' bytes.' . PHP_EOL;
    return $output;
}

function serializeStringOld(string $string): string {
    // Test the performance of the Ntriples::escapeString() method.
    $start = microtime(true);

    $ntriples = new NtriplesOld();
    $output = $ntriples->serialiseValue(['type' => 'literal', 'value' => $string]);

    $end = microtime(true);
    echo 'NtriplesOld::serialise() took ' . ($end - $start) . ' seconds. Memory usage: ' . memory_get_usage() . ' bytes. Memory peak usage: ' . memory_get_peak_usage() . ' bytes.' . PHP_EOL;
    return $output;
}

function test_individual_characters(): void {
    foreach (range("\x00", "\xFF") as $char) {
        $test_string = $char;
        echo "Character to serialize: $test_string" . PHP_EOL;
        // Test the performance of the Ntriples::serialise() method.
        $serialized_string = serializeString($test_string);
        $serialized_string_old = serializeStringOld($test_string);
        // Print the initial and the escaped string.
        echo 'Number of characters: ' . strlen($test_string) . PHP_EOL;
        echo 'Initial: ' . $test_string . PHP_EOL;
        echo 'Escaped: ' . $serialized_string . PHP_EOL;
        echo 'Escaped old: ' . $serialized_string_old . PHP_EOL;

        if ($serialized_string !== $serialized_string_old) {
            echo "'$serialized_string' !== '$serialized_string_old'";
            break;
        }
    }
}

$test_string = '';
foreach (range("\x00", "\xFF") as $char) {
    $test_string .= $char;
}

$test_string .= '位於';

// Test the performance of the Ntriples::serialise() method.
$serialized_string = serializeString($test_string);
$serialized_string_old = serializeStringOld($test_string);
echo 'Number of characters: ' . strlen($test_string) . PHP_EOL;
// Print the initial and the escaped string but only the first 100 characters.
echo 'Initial: ' . PHP_EOL . $test_string . PHP_EOL;
echo 'Escaped: ' . PHP_EOL . $serialized_string . PHP_EOL;
echo 'Escaped old: ' . PHP_EOL . $serialized_string_old . PHP_EOL;

$some_very_long_string = serializeString(str_repeat($test_string, 100));
$some_very_long_string_old = serializeStringOld(str_repeat($test_string, 100));
echo 'Number of characters: ' . strlen(str_repeat($test_string, 100)) . PHP_EOL;

$some_very_long_string = serializeString(str_repeat($test_string, 1000));
$some_very_long_string_old = serializeStringOld(str_repeat($test_string, 1000));
echo 'Number of characters: ' . strlen(str_repeat($test_string, 1000)) . PHP_EOL;

$some_very_long_string = serializeString(str_repeat($test_string . str_repeat('abcdefghigklmnopqrstuvwxyz', 50), 100));
$some_very_long_string_old = serializeStringOld(str_repeat($test_string . str_repeat('abcdefghigklmnopqrstuvwxyz', 50), 100));
echo 'Number of characters: ' . strlen(str_repeat($test_string . str_repeat('abcdefghigklmnopqrstuvwxyz', 50), 100)) . PHP_EOL;

// Print only the first 100 characters of the string.
echo 'Initial: ' . substr($some_very_long_string, 0, 100) . PHP_EOL;
echo 'Escaped: ' . substr($some_very_long_string, 0, 100) . PHP_EOL;
echo 'Escaped old: ' . substr($some_very_long_string_old, 0, 100) . PHP_EOL;


```