@idimopoulos 2023-03-22
**Disclaimer**: According to the documentation, 4-byte and 8-byte characters, which are sequences of other characters, can still interfere. For example, if 2 4-byte characters can form an 8-byte character, and they happen to be in sequence, these characters can be wrongly interpreted as an 8-byte character. But it is very hard for me to identify these and check them and make sure it is supported in any encodings (?). I do not have a case or report on this but I guess I will go with the 99%.