@claudiu-cristea 2023-03-23
> I wonder why Github Actions don't run here. Any idea? [Config file](https://github.com/sweetrdf/easyrdf/blob/master/.github/workflows/Tests.yml?rgh-link-date=2023-03-23T13%3A03%3A36Z) should be fine though.

Did you check the settings?