@idimopoulos 2023-03-23
I updated the branch but it is still failing. I also ran it on my repo to check the results. It seems that the 
```
php vendor/bin/php-cs-fixer fix --verbose --dry-run --format=checkstyle
```
is failing but it is not failing locally. Plus, the tests are green.