@zozlak 2023-03-24
Sorry for being late to the party ;) if you want to push more you can note that:

* The [n-triples specification](https://www.w3.org/TR/n-triples/#n-triples-grammar) allows almost all characters in the U+0 to U+10FFFF UNICODE space to just stay as they are. Moreover, the forbidden characters are all single byte so they can be very effectively escaped in one call to [strtr()](https://www.php.net/manual/en/function.strtr.php). Something like [here](https://github.com/sweetrdf/rdfHelpers/blob/master/src/rdfHelpers/NtriplesUtil.php#L106-L112) (just the linked implementation neglects existence of characters above U+10FFFF but [your code takes care of them already](https://github.com/sweetrdf/easyrdf/pull/25/files#diff-fec1dc558f11c5be9139ce296d590927743b430bc965e175c35e68c67325aac7R110-R134) in probably the best possible way).
* It's a pure waste of time to create the forbidden characters array every time the function is called. It should be just a constant or static property.