@idimopoulos 2023-03-24
> * It's a pure waste of time to create the forbidden characters array every time the function is called. It should be just a constant or static property.

@zozlak , thanks for the review. Initially I wanted to do this but the problem is that the control characters have weird pasting so it is not easy to build an array like
```php
array(
 '' => '\0',
 '' => '\1'
```
etc, as it will confuse some users or might accidentally be encoded by some IDEs, commits or whatever. That is why I am constructing the array.

I could set it in a local variable and have it constructed once per object and only on demand though. I will prepare the commit and let you decide on that.