@idimopoulos 2023-03-27
> Thank you @idimopoulos for this pull request! Also thanks @claudiu-cristea for your fast feedback.
> 
> These performance improvements look great!
> 
> You said `Almost all functionality kept`, I am wondering which points do we need to discuss further? As far as I understand your text, you say that the serializer at some point creates the wrong serialization for a given sequence. Is that correct?
> 
> > Please, correct me if I am wrong, but I can assume that the current implementation, can confuse sequences for multibytes if it randomly happens. If I am correct, this is a bug I had to fix here too.
> 
> The part about `randomly happens` I don't understand. Do you mean the ordering or is it in relation to something? Does it make sense to create a test which runs x times and uses a random generated string each time? This way we could check its behavior to some extend.
> 
> ~I wonder why Github Actions don't run here. Any idea? [Config file](https://github.com/sweetrdf/easyrdf/blob/master/.github/workflows/Tests.yml) should be fine though.~ Solved with #26

Wow... completely missed this :( sorry....
> You said `Almost all functionality kept`, I am wondering which points do we need to discuss further? As far as I understand your text, you say that the serializer at some point creates the wrong serialization for a given sequence. Is that correct?
Yes. Mainly, since it sequencially looks for multibytes, it might further push the text to wrongly assume a character. The test proves it too. For example, if we had a text 'abcd' but for multibyte characters, 'bc' ascii codes match the Euro symbol, the serializer would translate this to 'a€d' disregarding the possibility that this is simply an 'abcd' text. It happens often with characters from 127 to 255.

> The part about `randomly happens` I don't understand. Do you mean the ordering or is it in relation to something? Does it make sense to create a test which runs x times and uses a random generated string each time? This way we could check its behavior to some extend.
'Randomly' might not be a correct word. I meant that in the sequence of characters, from 0 to 10FFFF, it might be that we did not check all possible combinations that might result into a multibyte character. They are not random, they are very specific. The current serialization trying to serialize characters from 0 to 255 as I wrote in the description kinda proves that. I used a loop to generate the characters from 0 to 255 and tried to serialize it. Part of the result is 
```
\u00C1\u00C3\u014C\u00AD\u00AE\u00AF\u00B0\u00B1\u00B2\u00B3\u00B4\u00B5\u00B6\u00B7\u00B8\u00B9\u00BA\u00BB\u00BC\u00BD\u00BE\u00BF\u00C0\u00C1\u00C3\u0145\u01C7\u0249\u02CB\u034D\u03CF\u0451\u04D3\u0555\u05D7\u0659\u06DB\u075D\u07DF\u18A2\u4965\u7A28\uAAEB\uDBAE\U00010C71\U0C\u00AD\u00AE
```
While this should start from `\u00C1` and go up 1 hex digit at a time, you can see that it changes swiftly for example to `\u014C` and then back to `\u00AD`. Then, a little later, it even detects the multibyte character `\U00010C71` which is not part of the 255 first characters for sure. I don't think that should be meant to happen and in order for my code to work properly, I had to fix that too.

You can verify that this is the way the previous serializer works by running the new test about missmatching characters but using the old code. 
This is the almost all functionality that I didn't keep.