@zozlak 2023-03-27
> Yes. Mainly, since it sequencially looks for multibytes, it might further push the text to wrongly assume a character. The test proves it too. For example, if we had a text 'abcd' but for multibyte characters, 'bc' ascii codes match the Euro symbol, the serializer would translate this to 'a€d' disregarding the possibility that this is simply an 'abcd' text. It happens often with characters from 127 to 255.

If your input is in UTF-8 it's impossible. If your input is in a code page then you should realize UTF was introduced over 30 years ago to solve your problems :-D